[storage] Add serialization for String #222 (#225)
* Remove unused casts

* Update serialize.rs

* Update serialize.rs

* cleanup some dead code attributes

+++ src/storage/write_ahead_log.rs
@@ -5,12 +5,10 @@ use std::io::Read;
 use std::io::Seek;
 use std::io::Write;
 
-#[allow(dead_code)]
 pub(crate) struct WriteAheadLog {
     file: std::fs::File,
 }
 
-#[allow(dead_code)]
 impl WriteAheadLog {
     pub(crate) fn clear(&mut self) -> Result<(), DbError> {
         Ok(self.file.set_len(0)?)
@@ -45,8 +43,8 @@ impl WriteAheadLog {
     }
 
     fn read_record(file: &mut std::fs::File) -> Result<WriteAheadLogRecord, DbError> {
-        let position = u64::deserialize(&Self::read_exact(file, u64::serialized_size() as u64)?)?;
-        let size = u64::deserialize(&Self::read_exact(file, u64::serialized_size() as u64)?)?;
+        let position = u64::deserialize(&Self::read_exact(file, u64::serialized_size())?)?;
+        let size = u64::deserialize(&Self::read_exact(file, u64::serialized_size())?)?;
 
         Ok(WriteAheadLogRecord {
             position,
